Add benchmarking and CI check for import time (#10596)
## Describe your changes

Adds a benchmark test to keep track of the `streamlit` import time and a
threshold check to ensure we are are not adding anything that would
cause the import time to increase significantly.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.