Enable better exception logging if `rich` is installed (#10650)
## Describe your changes

We recently removed rich as a required dependency in
https://github.com/streamlit/streamlit/pull/10320. This PR automatically
enable better exception logging if `rich` is installed in the local
environment, instead of also requiring the config option to be
configured. This is mainly targeted for community cloud to simplify
their configuration, but might also be useful for some users.

## Testing Plan

- Updated tests to validate that `rich` is activated if `rich` is
installed (which is the case in our testing environment)

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.